pyproject.toml
[project]
name = "tdcollector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.3",
    "typer>=0.15.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
